+++ src/main/java/de/alpharogroup/db/entity/nameable/activatable/ActivatableNameEntity.java
@@ -22,23 +22,24 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.db.entity.visibility;
+package de.alpharogroup.db.entity.nameable.activatable;
 
 import java.io.Serializable;
 
 import javax.persistence.Column;
 import javax.persistence.MappedSuperclass;
 
-import de.alpharogroup.db.entity.base.SequenceBaseEntity;
-import lombok.AllArgsConstructor;
+import de.alpharogroup.db.entity.nameable.NameEntity;
+import de.alpharogroup.db.entity.nameable.IdentifiableNameableActivatable;
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
 /**
- * The class {@link VisibilityEntity} is a base entity with a flag 'visible' that indicates if the
- * entity is visible or not.
+ * The class {@link ActivatableNameEntity} is a base entity for a table with a single value
  *
  * @param <PK>
  *            the generic type of the id
@@ -47,17 +48,28 @@ import lombok.experimental.SuperBuilder;
 @Getter
 @Setter
 @NoArgsConstructor
-@AllArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
 @SuperBuilder
-public abstract class VisibilityEntity<PK extends Serializable> extends SequenceBaseEntity<PK>
-	implements
-		IdentifiableVisibility<PK>
+public abstract class ActivatableNameEntity<PK extends Serializable> extends NameEntity<PK>
+	implements IdentifiableNameableActivatable<PK>
 {
 
 	/** The serial Version UID. */
 	private static final long serialVersionUID = 1L;
 
-	/** The attribute visible, if true this entity is visible. */
-	@Column(name = "visible")
-	private boolean visible;
+	/** The attribute active indicates if this entity is active. */
+	@Column(name = "active")
+	boolean active;
+
+	/**
+	 * Instantiates a new {@link ActivatableNameEntity} with the given name
+	 *
+	 * @param name
+	 *            the name
+	 */
+	public ActivatableNameEntity(String name)
+	{
+		super(name);
+	}
+
 }