+++ src/main/java/de/alpharogroup/db/entity/nameable/versionable/VersionableNameUUIDEntity.java
@@ -22,42 +22,49 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.db.entity.visibility;
+package de.alpharogroup.db.entity.nameable.versionable;
 
-import java.io.Serializable;
-
-import javax.persistence.Column;
 import javax.persistence.MappedSuperclass;
+import javax.persistence.Version;
 
-import de.alpharogroup.db.entity.base.SequenceBaseEntity;
-import lombok.AllArgsConstructor;
+import de.alpharogroup.db.entity.nameable.NameUUIDEntity;
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
 /**
- * The class {@link VisibilityEntity} is a base entity with a flag 'visible' that indicates if the
- * entity is visible or not.
- *
- * @param <PK>
- *            the generic type of the id
+ * The class {@link VersionableNameUUIDEntity} is a base entity for a table with a single value
  */
 @MappedSuperclass
 @Getter
 @Setter
 @NoArgsConstructor
-@AllArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
 @SuperBuilder
-public abstract class VisibilityEntity<PK extends Serializable> extends SequenceBaseEntity<PK>
-	implements
-		IdentifiableVisibility<PK>
+public abstract class VersionableNameUUIDEntity extends NameUUIDEntity
 {
 
 	/** The serial Version UID. */
 	private static final long serialVersionUID = 1L;
 
-	/** The attribute visible, if true this entity is visible. */
-	@Column(name = "visible")
-	private boolean visible;
+	/**
+	 * The version property for the optimistic lock value
+	 **/
+	@Version
+	Integer version;
+
+	/**
+	 * Instantiates a new {@link VersionableNameUUIDEntity} with the given name
+	 *
+	 * @param name
+	 *            the name
+	 */
+	public VersionableNameUUIDEntity(String name)
+	{
+		super(name);
+	}
+
 }